+++ packages/manager-tools/manager-static-analysis-kit/src/adapters/eslint/rules/eslint-naming-conventions.ts
@@ -1,4 +1,4 @@
-import { ESLint, Linter } from 'eslint';
+import { Linter } from 'eslint';
 import checkFile from 'eslint-plugin-check-file';
 
 import {
@@ -11,7 +11,7 @@ import {
 export const checkFileEslintConfig: Linter.FlatConfig[] = [
   {
     plugins: {
-      'check-file': checkFile as unknown as ESLint.Plugin,
+      'check-file': checkFile,
     },
     rules: {
       'check-file/no-index': 'error',