+++ packages/manager/apps/pci-workflow/eslint.config.mjs
@@ -9,11 +9,9 @@ import {
   complexityTsJsConfig,
 } from '@ovh-ux/manager-static-analysis-kit/eslint/complexity';
 import { htmlEslintConfig } from '@ovh-ux/manager-static-analysis-kit/eslint/html';
-import { cssEslintConfig } from '@ovh-ux/manager-static-analysis-kit/eslint/css';
 import { tailwindJsxConfig } from '@ovh-ux/manager-static-analysis-kit/eslint/tailwind-jsx';
 import { tanStackQueryEslintConfig } from '@ovh-ux/manager-static-analysis-kit/eslint/tanstack';
 import { vitestEslintConfig } from '@ovh-ux/manager-static-analysis-kit/eslint/tests';
-import { tailwindSyntax } from '@eslint/css/syntax';
 
 export default [
   javascriptEslintConfig,
@@ -45,14 +43,6 @@ export default [
   prettierEslintConfig,
   complexityJsxTsxConfig,
   complexityTsJsConfig,
-  {
-    ...cssEslintConfig,
-    files: ['**/*.css', '**/*.scss'],
-    languageOptions: {
-      ...cssEslintConfig.languageOptions,
-      customSyntax: tailwindSyntax,
-    },
-  },
   {
     files: ['**/*.spec.*'],
     rules: {