+++ packages/manager-tools/manager-static-analysis-kit/CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.15.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-static-analysis-kit@0.14.0...@ovh-ux/manager-static-analysis-kit@0.15.0) (2025-12-05)
+
+
+### Features
+
+* use fixed versions for tests setup ([f917f0c](https://github.com/ovh/manager/commit/f917f0cb6a3c06f1cceb8a1bca8d1f42d66c4ee7)), closes [#MANAGER-20363](https://github.com/ovh/manager/issues/MANAGER-20363)
+
+
+
+
+
+# [0.14.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-static-analysis-kit@0.13.0...@ovh-ux/manager-static-analysis-kit@0.14.0) (2025-11-27)
+
+
+### Features
+
+* extends forge cli to create module and static kit update ([992f693](https://github.com/ovh/manager/commit/992f693fdd17ba82f2afb8f6cdfeb8fbd063929c))
+
+
+
+
+
+# [0.13.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-static-analysis-kit@0.12.0...@ovh-ux/manager-static-analysis-kit@0.13.0) (2025-11-16)
+
+
+### Features
+
+* pnpm hybrid improvements ([3b45577](https://github.com/ovh/manager/commit/3b455770da0c7aeb7603f7c731abf6fac9eb4b3f))
+
+
+
+
+
+# [0.12.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-static-analysis-kit@0.11.2...@ovh-ux/manager-static-analysis-kit@0.12.0) (2025-11-02)
+
+
+### Features
+
+* **muk:** static quality improvement ([2ca8694](https://github.com/ovh/manager/commit/2ca8694747cce2ea2901b6d04286ac07a65be03d))
+
+
+
+
+
+## [0.11.2](https://github.com/ovh/manager/compare/@ovh-ux/manager-static-analysis-kit@0.11.1...@ovh-ux/manager-static-analysis-kit@0.11.2) (2025-10-23)
+
+
+### Bug Fixes
+
+* static lint issue ([465deb1](https://github.com/ovh/manager/commit/465deb11d7203fd5b0252c1a73474e4b24476359)), closes [#MANAGER-19481](https://github.com/ovh/manager/issues/MANAGER-19481)
+
+
+
+
+
 ## [0.11.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-static-analysis-kit@0.11.0...@ovh-ux/manager-static-analysis-kit@0.11.1) (2025-10-12)
 
 **Note:** Version bump only for package @ovh-ux/manager-static-analysis-kit